+++ 0x01-python_async_function/4-tasks.py
@@ -7,7 +7,7 @@ from typing import List
 task_wait_random = __import__('3-tasks').task_wait_random
 
 
-async def task_wait_n(n: int, max_delay: int) -> List[int]:
+async def task_wait_n(n: int, max_delay: int) -> List[float]:
     """Spawns wait_random n times with the specified max_delay.
     """
     delays = []
